+++ TopModel.Generator.Javascript/GeneratorRegistration.cs
@@ -10,7 +10,23 @@ public class GeneratorRegistration : IGeneratorRegistration<JavascriptConfig>
     /// <inheritdoc cref="IGeneratorRegistration{T}.Register" />
     public void Register(IServiceCollection services, JavascriptConfig config, int number)
     {
-        TrimSlashes(config, c => c.ApiClientFilePath);
+        if (config.ApiClientFilePath != null)
+        {
+            TrimSlashes(config, c => c.ApiClientFilePath);
+            if (!config.ApiClientFilePath.Contains("{fileName}"))
+            {
+                config.ApiClientFilePath = config.ApiClientFilePath + "/{fileName}";
+            }
+        }
+        else if (config.ApiMode == TargetFramework.ANGULAR)
+        {
+            config.ApiClientFilePath = "{module}/{fileName}.service";
+        }
+        else
+        {
+            config.ApiClientFilePath = "{module}/{fileName}";
+        }
+
         TrimSlashes(config, c => c.ApiClientRootPath);
         TrimSlashes(config, c => c.DomainPath);
         TrimSlashes(config, c => c.FetchPath);